dom/tests/mochitest/mouselock/file_fullscreen.html

<!DOCTYPE HTML>
<html>
  <!-- https://bugzilla.mozilla.org/show_bug.cgi?id=92264 -->
  
  <head>
    <title>
      Test for Bug 633602
    </title>
    <script src="/tests/SimpleTest/EventUtils.js">
    </script>
    <script src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script type="application/javascript" src="mouselock_util.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"
    />
  </head>
  
  <body onload="runTests();">
    <a target="_blank"
      href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
        Mozilla Bug 633602
    </a>
    <p id="display">
    </p>
    <div id="content" style="display: none">
    </div>
    <canvas id="canvas" width="150" height="150">
    </canvas>
    <pre id="test">
      <script>
        SimpleTest.waitForExplicitFinish();

        const pointer = navigator.pointer;
        const canvas = document.getElementById("canvas");
        
        var lockTestDone = false,
          unlockTestDone = false,
          finish = false;

        document.addEventListener("mozfullscreenchange", function(e) {
          if (document.mozFullScreen) {
            if (document.mozFullScreenElement === canvas) {
              if (lockTestDone === true && unlockTestDone === true) {
                pointer.lock(
                  canvas,
                  function() {
                    ok(pointer.islocked(), "Pointer locked successfully \
                      after attaining fullscreen mode");

                    finish = true;
                    document.mozCancelFullScreen();
                  },
                  function() {
                    ok(false, "Mouse could not achieve lock when it \
                      should have, element had fullscreen mode");
                  }
                );
              } else {
                pointer.lock(
                  canvas,
                  function() {
                    ok(pointer.islocked(), "Mouse should be locked when \
                      it gets a DOM element that is fullscreened");
                    pointer.unlock();

                    pointer.lock(
                      canvas,
                      function() {
                        ok(pointer.islocked(), "Mouse should be locked \
                          when the success callback is fired");
                        lockTestDone = true;
                        document.mozCancelFullScreen();
                      },
                      function() {
                        ok(false, "Mouse failed to lock after unlock");
                      }
                    );
                  },
                  function() {
                    ok(false, "Mouse could not achieve lock when \
                      it should have, fullscreen was available");
                  }
                );
              }
            } else {
              pointer.lock(
                canvas,
                function() {
                  ok(false, "Mouse locked when it should not have, \
                     locked element is not in fullscreen mode");
                },
                function() {
                  isnot(pointer.islocked(), true, "Mouse should only \
                    be locked if the locked element is in fullscreen mode");
                  unlockTestDone = true;

                  document.mozCancelFullScreen();
                }
              );
            }
          } else if (finish === true) {
            isnot(pointer.islocked(), true, "Mouse should unlock when \
              fullscreen mode is taken away from element");

            SimpleTest.finish();
          } else {
            if (lockTestDone === true && unlockTestDone === false) {
              pointer.lock(
                canvas,
                function() {
                  ok(false, "Mouse locked when it should not have, not in \
                     fullscreen");
                },
                function() {
                  isnot(pointer.islocked(), true, "Mouse should be unlocked \
                    if a failure callback is fired.");

                  document.body.mozRequestFullScreen();
                }
              );
            } else {
              canvas.mozRequestFullScreen();
            }
          }
        }, false);

        function runTests() {
          pointer.lock(
            canvas,
            function() {
              ok(false, "Pointer locked when it should not have, element \
                isn't in fullscreen mode")
            },
            function() {
              isnot(pointer.islocked(), true, "Mouse can't be locked if \
                an element isn't in fullscreen mode");

              canvas.mozRequestFullScreen();
            }
          );
        };
      </script>
    </pre>
  </body>

</html>
